+++ hudi-spark-datasource/hudi-spark-common/src/main/scala/org/apache/hudi/HoodieMergeOnReadRDD.scala
@@ -22,7 +22,7 @@ import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.mapred.JobConf
 
 import org.apache.hudi.HoodieBaseRelation.{BaseFileReader, projectReader}
-import org.apache.hudi.LogIteratorUtils.CONFIG_INSTANTIATION_LOCK
+import org.apache.hudi.LogFileIterator.CONFIG_INSTANTIATION_LOCK
 import org.apache.hudi.common.model.OverwriteWithLatestAvroPayload
 import org.apache.hudi.exception.HoodieException
 import org.apache.hudi.hadoop.utils.HoodieRealtimeRecordReaderUtils.getMaxCompactionMemoryInBytes
